feat: annotate FetchRobotsTxt function as failable
This commit set the return type of the `FetchRobotsTxt` to
error to indicate that it can potentially fail due to network IO.
In addition, error handling blocks in the function are updated to
propagate the received error when it is encountered. Lastly, the
provision method is modified to run `FetchRobotsTxt` synchrously
with a timeout of 10 seconds set for the darkvisitors API query rather than
as a goroutine. This change is made because there is no way to wait
on a variable in the request context, so while using a goroutine for
the query avoids blocking the module startup, it can lead to difficult
to debug bugs if the robots.txt query does not return before the middleware
is triggered. Running the query synchrously during the provision phase
with a set timeout is a compromise that eliminate this issue.